Move utils into polaffini
