+++ scripts/train.py
@@ -10,7 +10,7 @@ import dwarp
 import SimpleITK as sitk
 import argparse
 import generators
-import utils
+import polaffini.utils as utils
 
 print("Num GPUs Available: ", len(tf.config.list_physical_devices('GPU')))
